# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão inteligente de fotografias e eventos caninos do PS (Photo Storage).",
        "title": "PS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte PS",
            "email": "suporte@ps.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/photographers": {
            "get": {
                "description": "Retorna a lista de fotógrafos do tenant autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photographers"
                ],
                "summary": "Listar fotógrafos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/photographer.Photographer"
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Cadastra um novo fotógrafo para o tenant autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photographers"
                ],
                "summary": "Criar fotógrafo",
                "parameters": [
                    {
                        "description": "Dados do Fotógrafo",
                        "name": "photographer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photographer.Photographer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/photographer.Photographer"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/photographers/{id}": {
            "get": {
                "description": "Retorna os detalhes de um fotógrafo específico pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photographers"
                ],
                "summary": "Obter fotógrafo por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Fotógrafo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/photographer.Photographer"
                        }
                    },
                    "400": {
                        "description": "ID obrigatório",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Fotógrafo não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de um fotógrafo existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photographers"
                ],
                "summary": "Atualizar fotógrafo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Fotógrafo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do Fotógrafo",
                        "name": "photographer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photographer.Photographer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/photographer.Photographer"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um fotógrafo do tenant autenticado",
                "tags": [
                    "Photographers"
                ],
                "summary": "Excluir fotógrafo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Fotógrafo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluído com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID obrigatório",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "photographer.Photographer": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer \" seguido do token JWT obtido no login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}